Feature #10762
Define a new object in SilverpeasConfigurationProperties: Context.
This object defines a configuration context that can be used by any components
implied in a task of SilverSetup as well by any scripts executed by those
tasks. The context is publicly available in the settings global variable and it
is represented by a Map. Once the build is finished, the properties set in the
context are then persisted into the file SILVERPEAS_HOME/configuration/.config.

The goal of the Context object is to allow any scripts and components used in
tasks to keep track of any configuration, migration or installation context so
that they can adapt their behaviour at the next execution.