Bug #10769
Restructure the different sections of settings of the plugin SilverSetup in
order to separate correctly what are for configuration of Silverpeas, what
are for installation and what are for the migration of the data source.

Add a new plugin setting property: timeout. It sets the timeout of the
waiting for JBoss to answer to the requests sent by the plugin. By
default it is set to 5mn (instead of 2mn) but it can be changed by
setting the property
silversetup {
  timeout = <the number of milliseconds to wait for>
}
in the SILVERPEAS_HOME/bin/build.gradle file.