Bug #10969
Fix the development mode checking.
Fix the command status parsing with CLI scripts (in JBossServer)

Add a way to specify a default value in variable placeholders in an expression.
For example, a variable without default value can be used as before in an
expression: ${MY_VARIABLE}.
Beside that, a default value can be indicated as such: ${MY_VARIABLE:true}
the semi-colon is used as the separator between the variable name and its
default value. When the variable placeholder is encountered within an
expression and it has no values set (for example in the config.properties),
then the default value is used to replace the variable placehodler.