Feature #12866
Clean up all the code and tests on resources that aren't anymore
available (Castor XML parsing of workflow process instances).

Update the dependencies.

Take into account the JCR repository is now managed by Jackrabbit Oak.
For doing, the JcrRepositoryFactory is updated and a new class,
JcrService is provided to give access to the JCR repository during the
configuration/migration phases.